refactor dotenv setup out of the entrypoint module
Some modules (routes/api.ts) may rely on the dotenv setup. This causes
issues when the setup logic is directly in the entrypoint module because
imports are hoisted, and so the dependent logic runs before the setup.
By moving the setup logic into its own module and importing it into the
entrypoint, we can ensure that the setup happens first.